requirements.txt
python-telegram-bot==21.6
yt-dlp>=2025.1.1
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0